feat(buttons): Add TerminalNavButton for enhanced navigation with hover effects
